Split up history tables into separate files
