+++ simbelmyne/src/main.rs
@@ -12,7 +12,7 @@ mod transpositions;
 mod move_picker;
 mod time_control;
 mod tests;
-mod search_tables;
+mod history_tables;
 
 #[derive(Parser)]
 #[command(author = "Sam Roelants", version = "0.1", about = "A simple perft tool.", long_about = None)]
