+++ simbelmyne/src/uci.rs
@@ -18,6 +18,7 @@ use uci::options::OptionType;
 use uci::options::UciOption;
 use crate::evaluate::pretty_print::print_eval;
 use crate::evaluate::Score;
+use crate::history_tables::history::HistoryTable;
 use crate::search::params::ASPIRATION_BASE_WINDOW;
 use crate::search::params::ASPIRATION_MAX_WINDOW;
 use crate::search::params::ASPIRATION_MIN_DEPTH;
@@ -37,7 +38,6 @@ use crate::search::params::RFP_MARGIN;
 use crate::search::params::RFP_THRESHOLD;
 use chess::perft::perft_divide;
 use crate::search::params::SearchParams;
-use crate::search_tables::HistoryTable;
 use crate::time_control::TimeController;
 use crate::time_control::TimeControlHandle;
 use crate::transpositions::TTable;
